requirements.txt
# Core
openai>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Agent Framework
crewai>=0.80.0

# Vector Store
chromadb>=0.5.0

# Web UI
streamlit>=1.40.0

# Data Sources
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
yfinance>=0.2.50

# PDF Generation
weasyprint>=62.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
responses>=0.25.0
